Swap view components to Svelte
In this experiment, the 2 view components implemented in Angular are
removed and replaced with roughly equivalent view components implemented
in Svelte.

Problems remaining:

Svelte web components always use shadow DOM, hiding them from the global
stylesheet; it appears that good solutions for styling and application
composed of many isolated shadow DOMs, are still very much a topic of
discussion and exploration. It should be possible to extract out just a
small slice of the global stylesheet needed for each view component, and
although that would still involve a lot of CSS duplication, it should be
manageable.

There is a VSCode plug-in, but it does not get along especially well
with the various files set up here for the Angular app. There are
various errors shown in Svelte component code, TypeScript errors,
although Svelte does not yet support using TypeScript in components.

In this application at least, the web components are initially created
without their properties set (or perhaps that the undefined), because of
asynchronous data loading perhaps. Svelte warns about this even though
it is okay in this case.